finish up hackers with xeditable and toastrs

+++ public/js/controllers/HackerCtrl.js
@@ -1,9 +1,9 @@
 angular.module('sabio.hacker.controller', ['sabio.hacker.service'])
   .controller('hackerController', HackerController)
 
-HackerController.$inject = ['hackerService', '$stateParams','$alertService']
+HackerController.$inject = ['hackerService', '$stateParams', '$alertService']
 
-function HackerController(hackerService, $stateParams, $alertService) { 
+function HackerController(hackerService, $stateParams, $alertService) {
   'use strict'
   var vm = this
   vm.tagline = 'Hack The Planet!'
@@ -11,33 +11,71 @@ function HackerController(hackerService, $stateParams, $alertService) {
   vm.formData = {}
   vm.hackers = []
 
-  if ($stateParams.id) {
-    hackerService.get($stateParams.id, onGetSuccess)
-  }
-  else {
-    hackerService.getAll(getAllSuccess, onError)
+  vm.checkName = _checkName;
+  vm.submitRow = _submitRow;
+  vm.removeRow = _removeRow;
+  vm.addRow = _addRow;
+
+  _initialize();
+
+  function _initialize() {
+    if ($stateParams.id) {
+      hackerService.get($stateParams.id, onGetSuccess)
+    }
+    else {
+      hackerService.getAll(getAllSuccess, onError)
+    }
   }
 
-  vm.insert = () => {
-    hackerService.post(vm.formData, onInsertSuccess, onError)
+  //  this is just a silly example but it shows how you can validate data before it gets submitted.
+  //   you could do an ajax call here to check if an email address is unique, for example.
+  function _checkName(data, id) {
+    if (!data)
+      return "Hacker name is required";
+
+    if (data.toLowerCase().indexOf("sabio") > -1)
+      return "Thou shalt not take Sabio's name in vain!!";
   }
-  vm.update = () => {
-    hackerService.put(vm.formData, onUpdateSuccess, onError)
+
+  function _addRow() {
+
+    vm.formData = { name: null };
+
+    vm.hackers.unshift(vm.formData);
   }
+
+  function _submitRow(data, id) {
+    if (id) {
+      hackerService.put(data, onUpdateSuccess, onError)
+    }
+    else {
+      hackerService.post(data, onInsertSuccess, onError)
+    }
+  }
+
+  function _removeRow(id, index) {
+    if (id) {
+      if (confirm("Are you sure you want to delete this hacker?")) {
+        hackerService.delete(id, onDeleteSuccess, onError)
+      }
+    }
+    else {
+      vm.hackers.splice(index, 1);
+    }
+
+
+  }
+
   vm.get = (id) => {
     hackerService.get(id, onGetSuccess, onError)
   }
-  vm.remove = (id) => {
-    hackerService.delete(id, onDeleteSuccess, onError)
-  }
-
 
   function onInsertSuccess(data) {
     vm.formData = null
-    if (data.item)
-      vm.hackers.push(data.item)
+    
+    vm.$alertService.success(data.item.name + " has been inserted");
 
-      vm.$alertService.success(data.item.name + " has been inserted");
+    _initialize();
   }
 
   function getAllSuccess(data) {
@@ -50,8 +88,10 @@ function HackerController(hackerService, $stateParams, $alertService) {
 
   function onUpdateSuccess(data) {
     vm.formData = null
-    if (data)
-      vm.hackers.push(data)
+    
+    vm.$alertService.success("Hacker was updated");
+
+    _initialize();
   }
 
   function onDeleteSuccess(data) {
@@ -65,12 +105,12 @@ function HackerController(hackerService, $stateParams, $alertService) {
   function onError(data) {
     var msg = "An error occured";
 
-    if(data.errors && data.errors.errmsg)
+    if (data.errors && data.errors.errmsg)
       msg += ": " + data.errors.errmsg;
 
-      vm.$alertService.error(msg, "Hacker error");
+    vm.$alertService.error(msg, "Hacker error");
 
-    console.log('Error:',data.errors);
+    console.log('Error:', data.errors);
   }
 
 }
